refactor(engine): remove dead surface and fix the composite truncation
Four things carried through every match without ever being read.

`PlayerData::effective_overall`, `TeamData::defense_rating`,
`attack_rating` and `goalkeeper_rating` had no callers outside their own
tests — three of the four composite team ratings never reached a
simulated match. `EventType::ShotOnTarget` was never emitted by any code
path, but both the report and the frontend carried a branch for it that
could not fire. `PlayerSnap` copied `stamina`, `strength` and
`leadership` on every action, twice, and nothing read any of them.

`position_attr_avg` built its composite with integer division and cast
back to `u8` before averaging, truncating every rating it produced. It
also collected into a `Vec` to count. Both are gone; the one caller left,
`midfield_rating`, feeds the possession contest.

The golden reports could not have caught that truncation, because the
compliance squads had every attribute set to the same number — so every
composite came out exact, and weighted selection was degenerate with all
candidates scoring identically. Attributes are now spread deterministically
around each player's rating, which is the third time a uniform or
defaulted test fixture has hidden a whole class of change from a golden.

`assign_ratings` cloned every player id and scanned linearly for each
one; it now borrows and hashes. It also missed players who had been
substituted off — they have a performance to rate even though they are
no longer in the eleven.

+++ src-tauri/crates/engine/src/live_match/helpers.rs
@@ -383,7 +383,7 @@ impl LiveMatchState {
     pub(super) fn effective_press(&self, pressing_side: Side) -> f64 {
         let team = self.team_ref(pressing_side);
         let base = team.position_attr_avg(Position::Midfielder, |p| {
-            ((p.stamina as u16 + p.tackling as u16 + p.pace as u16) / 3) as u8
+            (p.stamina as f64 + p.tackling as f64 + p.pace as f64) / 3.0
         });
         let modifier = play_style_modifier(team.play_style, PlayStylePhase::Press, true);
         base * modifier
